requirements.txt
amqp==2.6.0
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.5
Django==3.0.7
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
future==0.18.2
kombu==4.6.10
Pillow==2.2.1
PyJWT==1.7.1
pytz==2020.1
redis==3.5.3
sqlparse==0.3.1
vine==1.3.0
